Changed price_per_interval from int to float

+++ table_definitions/pricing_scheme.sql
@@ -1,7 +1,7 @@
 CREATE TABLE pricing_scheme(
 	payment_type varchar(10) NOT NULL,
     pass_required varchar(3),
-    price_per_interval int,
+    price_per_interval float,
     parking_interval date,
     PRIMARY KEY(payment_type) );
